fix: use helix-shared-tokencache
BREAKING CHANGE: tokencache moved to shared
