--- src/cache/FSCacheManager.d.ts
@@ -1,28 +1,0 @@
-/*
- * Copyright 2022 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-import { FSCachePlugin } from "./FSCachePlugin";
-
-export declare interface FSCacheManagerOptions {
-  log: Console;
-  dirPath: string;
-  type: string;
-}
-
-export declare class FSCacheManager {
-  constructor(opts: FSCacheManagerOptions);
-
-  listCacheKeys():Promise<string[]>;
-
-  hasCache(key:string):Promise<boolean>;
-
-  getCache(key:string):Promise<FSCachePlugin>;
-}
