--- test/cache/encrypt.test.js
@@ -1,27 +1,0 @@
-/*
- * Copyright 2021 Adobe. All rights reserved.
- * This file is licensed to you under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License. You may obtain a copy
- * of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under
- * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- * OF ANY KIND, either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-
-/* eslint-env mocha */
-import assert from 'assert';
-import { decrypt, encrypt } from '../../src/cache/encrypt.js';
-
-describe('Encrypt Tests', () => {
-  it('encrypts and decrypts data', async () => {
-    const text = 'Hello, world.';
-    const key = '12345678';
-
-    const encrypted = encrypt(key, Buffer.from(text, 'utf-8'));
-    const decrypted = decrypt(key, encrypted);
-
-    assert.strictEqual(decrypted.toString('utf-8'), text);
-  });
-});
