+++ test/dev/decrypt.js
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Adobe. All rights reserved.
+ * Copyright 2023 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,22 +10,20 @@
  * governing permissions and limitations under the License.
  */
 
-/**
- * @typedef {import("@azure/msal-node").ISerializableTokenCache} ISerializableTokenCache
- *
- * @extends TokenCacheContext
- */
-export class MockTokenCacheContext {
-  constructor(opts = {}) {
-    const context = this;
-    Object.assign(this, {
-      cacheHasChanged: false,
-      tokens: '',
-      /** @type ISerializableTokenCache */
-      tokenCache: {
-        serialize: () => context.tokens,
-        deserialize: (value) => { context.tokens = value; },
-      },
-    }, opts);
+/* eslint-disable no-console */
+import { decrypt } from '@adobe/helix-shared-tokencache';
+import fs from 'fs/promises';
+
+async function run() {
+  if (process.argv.length < 3) {
+    console.error('usage: node src/decrypt.js <file>');
+    return -1;
   }
+  const data = await fs.readFile(process.argv[2]);
+  const decrypted = decrypt('***', data);
+  process.stdout.write(JSON.stringify(JSON.parse(decrypted.toString('utf-8')), null, 2));
+  process.stdout.write('\n');
+  return 0;
 }
+
+run().then(process.exit);
